Split gulp setting task to env and plugins. Updated references.
