+++ .gitignore
@@ -5,7 +5,7 @@
 # Dependency directories
 dist/
 node_modules/
-src/common/*.properties.json
+src/environment
 src/consumer/public/manifest.json
 src/consumer/public/brand
 src/consumer/public/sw.js
@@ -14,4 +14,5 @@ src/consumer/next-sitemap.js
 src/**/lib
 src/**/.firebase
 **/firebase-debug.log
+**/ui-debug.log
 src/admin/public
