+++ scripts/templates/common.environment.js
