+++ scripts/templates/server.environment.js
